init commit, refactor gent to have split sections
